Change read.c to aware sexp abstraction.
